fix rename token
